Resolve "tools/ingest.py and tools/catalog.py need to know region of heliocloud instance"

Closes #169

See merge request heliocloud/platform!161

+++ registry/lambdas/app/ingest_lambda.py
@@ -11,9 +11,10 @@ from .aws_utils.s3 import get_manifest_from_s3
 from .ingest.ingester import Ingester
 
 
+# Ingester does not require context parameter, but AWS Lambda service still requires
+# context as a parameter so the handler has the right method signature.
+# pylint: disable=unused-argument
 def handler(event, context) -> dict:
-    # pylint: disable=unused-argument
-    # context argument is required by AWS lambda service
     """
     AWS Lambda handler for the Ingest service of HelioCloud.
 
@@ -38,45 +39,50 @@ def handler(event, context) -> dict:
         session=session, bucket_name=ingest_bucket, entry_key=entry_key
     )
 
-    # Get the manifest for each entry
-    results = []
+    # For each entry in the entries file, get the manifest and run an Ingester instance
+    updates = []
     s3_client = session.client("s3")
     for entry_ds in entry_ds_list:
         # join ingest folder (ex. my_job/), index split (ex. MMS/), and manifest.csv
         manifest_key = os.path.join(ingest_folder, entry_ds.index.rsplit("/", 1)[1], "manifest.csv")
 
+        # Get the manifest file
         manifest_df = get_manifest_from_s3(
             session=session, bucket_name=ingest_bucket, manifest_key=manifest_key
         )
 
-        ingester = Ingester(
-            session=session,
-            ingest_bucket=ingest_bucket,
-            ingest_folder=ingest_folder,
-            entry_dataset=entry_ds,
-            manifest_df=manifest_df,
-            ds_repo=get_dataset_repository(),
-        )
-        results.append(ingester.execute())
+        # Run the Ingester, catching any exceptions that arise
+        update = {"dataset": entry_ds.dataset_id, "num_files_updated": 0, "error": None}
+        # pylint: disable=broad-exception-caught
+        try:
+            result = Ingester(
+                session=session,
+                ingest_bucket=ingest_bucket,
+                ingest_folder=ingest_folder,
+                entry_dataset=entry_ds,
+                manifest_df=manifest_df,
+                ds_repo=get_dataset_repository(),
+            ).execute()
+            update["num_files_updated"] = result.files_contributed
+            s3_client.delete_object(Bucket=ingest_bucket, Key=manifest_key)
 
-        s3_client.delete_object(Bucket=ingest_bucket, Key=manifest_key)
+        # Ingester failed, so record the dataset impacted and the exception that occurred.
+        # so it can be reported back
+        except Exception as ex:
+            update["error"] = str(ex)
+        # pylint: enable=broad-exception-caught
 
-    # Remove the entries file
-    s3_client.delete_object(Bucket=ingest_bucket, Key=entry_key)
+        # Store a record of the update
+        updates.append(update)
+
+    # If all entries process successfully, remove the entries file
+    if len(updates) == len(entry_ds_list):
+        s3_client.delete_object(Bucket=ingest_bucket, Key=entry_key)
+
     s3_client.close()
 
-    # Information to return back in the Lambda response payload stream
-
-    updates = {
-        "datasets_updated": [],
-        "files_contributed": [],
-    }
-
-    for result in results:
-        updates["datasets_updated"].append(result.dataset_updated)
-        updates["files_contributed"].append(result.files_contributed)
-
-    return updates
+    # Return a dictionary of the results
+    return {"num_datasets_updated": len(updates), "updates": updates}
 
 
 # pylink: enable=unused-argument
